moved 'create_project_folders' and 'get_project_basic_paths'
