+++ openpype/modules/ftrack/event_handlers_user/action_create_project_structure.py
@@ -1,7 +1,10 @@
 import re
 
+from openpype.pipeline.project_folders import (
+    get_project_basic_paths,
+    create_project_folders,
+)
 from openpype_modules.ftrack.lib import BaseAction, statics_icon
-from openpype.api import get_project_basic_paths, create_project_folders
 
 
 class CreateProjectFolders(BaseAction):
@@ -81,7 +84,7 @@ class CreateProjectFolders(BaseAction):
                 }
 
             # Invoking OpenPype API to create the project folders
-            create_project_folders(basic_paths, project_name)
+            create_project_folders(project_name, basic_paths)
             self.create_ftrack_entities(basic_paths, project_entity)
 
             self.trigger_event(
